Initial test coverage and cosmetic changes
On first look at the code one would notice that the existing code had
no tests. This commit adds such tests with the aim to describe and
provide clarity on the behaviour of he existing code.

Also a few cosmetic issues like better method single/ double quotes,
indentation and addition of rspec coloring have been include within
this commit.

FYI: The idea of a #fixture_file_path is inspired by [1], in this
case it appears we may be requiring such as things progress.

[1] https://github.com/anarayanan86/rails/blob/master/activesupport/lib/active_support/testing/file_fixtures.rb#L22