Dryup: Move #coordinates_colours to Parser
Upon review it became clear that the #coordinates_colours method was
being repeated across wrapper classes and it main purpose is in line
with that of the parser, hence why.

For this to actually be a concrete move the a lot files have been
affected. Tests have also been written to properly cover
coordinates_colour